requirements.txt
Django==2.2.7
django-taggit==1.1.0
pytz==2019.3
sqlparse==0.3.0
